requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
gunicorn==23.0.0
google-cloud-bigquery==3.27.0
google-auth==2.37.0
pydantic==2.10.4
pydantic-settings==2.7.0